Implement Base, Enemy. Refactored and added functionality to Tower.
Tower now checks for its first target within range and switches successfully to the next enemy in line (second that entered its range).

Tower attacks enemies implemented (with enemy HP and tower damage).

Enemy deaths and bugs associated with it fixed. (Had to do with removeNode() which drops the mesh and messes with drawLine!)